requirements.txt
azure-core==1.21.1
azure-iot-device==2.9.0
azure-iot-hub==2.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
deprecation==2.1.0
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.0.0
idna==3.3
importlib-metadata==4.10.1
isodate==0.6.1
itsdangerous==2.0.1
janus==1.0.0
Jinja2==3.0.3
MarkupSafe==2.0.1
msrest==0.6.21
oauthlib==3.1.1
packaging==21.3
paho-mqtt==1.6.1
pyparsing==3.0.7
PySocks==1.7.1
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.0
requests-unixsocket==0.3.0
six==1.16.0
typing-extensions==4.0.1
uamqp==1.5.1
urllib3==1.26.8
visitor==0.1.3
Werkzeug==2.0.2
WTForms==3.0.1
zipp==3.7.0
